refactor: code refactor complete
