+++ middleware/multerMiddleware.js
@@ -1,0 +1,17 @@
+const multer = require('multer');
+const config = require('../config');
+const path = require('path');
+
+const storage = multer.diskStorage({
+  destination: function (req, file, cb) {
+    cb(null, config.UPLOAD_FOLDER);
+  },
+  filename: function (req, file, cb) {
+    cb(
+      null,
+      file.fieldname + '-' + Date.now() + path.extname(file.originalname)
+    );
+  },
+});
+
+module.exports = storage;
